requirement.txt
absl-py==2.3.1
altair==4.2.2
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
entrypoints==0.4
flatbuffers==25.9.23
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
imutils==0.5.4
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==1.23.5
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==3.19.6
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.4
wrapt==2.0.1
zipp==3.23.0

